backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
openai==1.98.0
pydantic==2.11.7
pydantic_core==2.33.2
httpx==0.28.1
